renamed upload_utility module to common_utility
