+++ common_utility/exceptions.py
