+++ common_utility/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "common_utility",
+  "version": "1.0.0",
+  "description": "",
+  "author": "SingularityNet",
+  "dependencies": {
+  	"serverless-offline": "5.12.0",
+    "serverless-plugin-tracing": "2.0.0",
+    "serverless-prune-plugin": "1.4.1",
+    "serverless-python-requirements": "5.0.0",
+    "singularitynet-platform-contracts": "0.3.4"
+
+  }
+}
