+++ common_utility/requirements.txt
@@ -1,0 +1,5 @@
+boto3
+alembic==1.3.1
+boto3==1.10.9
+PyMySQL==0.9.3
+SQLAlchemy==1.3.11
